+++ db/lou_seeds.sql
@@ -1,0 +1,11 @@
+-- id=1
+INSERT INTO `gameswap_db`.`locations` (`name`,`city`,`state`,`zip`) 
+VALUES ("University of Central Florida", "Orlando", "FL", 32816);
+
+-- id=2
+INSERT INTO `gameswap_db`.`locations` (`name`,`city`,`state`,`zip`) 
+VALUES ("Full Sail University", "Orlando", "FL", 32792);
+
+-- id=3
+INSERT INTO `gameswap_db`.`locations` (`name`,`city`,`state`,`zip`) 
+VALUES ("Rollins College", "Orlando", "FL", 32789);
